feat: replace container-interop
This patch updates the package to act as a replacement for
container-interop, by adding class_alias statements via an autoloader
that alias the various container-interop interfaces to the psr/container
equivalents.

This allows us to remove the functionality within the abstract plugin
manager that decorates the parent container as an interop-container if
it is a psr/container, and essentially allows creates a dual-inheritance
tree that allows either interface to be used at any time, regardless of
the typehint.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>