fix: alphabetize import statements in generated factories
Previously, the imports were _mostly_ alphabetized. However, with the
switch to Laminas, the container import was now happening incorrectly
with regards to the factory interface. Additionally, it makes sense to
alphabetize to ensure we follow our own CS.

This patch updates the factory generation code to alphabetize import
statements before injecting them in the template.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>