feat: replace container-interop by providing it directly
Per feedback on my original commit (7066cde01), this patch takes a
slightly different approach. Instead of using class aliases, it instead
provides the container-interop interfaces directly, importing them as
part of the package and providing autoloading for them.

All other changes remain the same, with one exception: because we are no
longer class aliasing, `ServiceLocatorInterface` MUST still implement
both PSR-11 and container-interop in order for typehints to continue to
work. However, the changes to `AbstractPluginManager` continue to be
valid, as the plugin manager only exercises the PSR-11 API. Since
container-interop at this version extends PSR-11, it also fulfills the
contract; instances can be used interchangeably internally, and there is
no need to perform any decoration. As such, I've also marked
`PsrContainerDecorator` deprecated, and slated it for removal in next
major.

Signed-off-by: Matthew Weier O'Phinney <matthew@weierophinney.net>