Use one sentence per line style
The intention of this change is to make the file easier to maintain,
rather than to have lines fit within a mandatory 72 - 80 char per/line
limit. Personally, this new approach makes content easier to move
around, to edit, and to remove as and when required, than arbitrary line
lengths that break sentences midway across one or more lines.

Signed-off-by: Matthew Setter <matthew@matthewsetter.com>