qa: re-add platform requirement exception for PHP 8.1
`ocramius/proxy-manager` is not supporting PHP 8.1 as of now. We could replace the dependency with `friendsofphp/proxy-manager-lts` or simply ignore platform requirement for PHP 8.1. For PHP 8.2, we should remove the exception for PHP 8.1 and if there are problems with the proxy manager, we should replace it with the LTS version.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>