qa: remove `MixedAssignment` suppressions as these are now suppressed by `psalm.xml`
Due to the way, psalm moves forward, this issue will be removed - most probably with v6.
The reasoning behind this is, that `mixed` is not bad per-se, depending on how it is used.
If a codebase accepts `mixed` in a method (even tho that might be revealing method complexity), that is a absolutely okay and thus, passing `mixed` and esp. assigning `mixed` is absolutely fine in case the following code is accepting mixed and/or the value is narrowed down at some point.

Signed-off-by: Maximilian Bösing <2189546+boesing@users.noreply.github.com>